Terminada función 2.A-editTypeOfFoodInRecipe y cambiado de lugar la función loadGetHighestRatedRecipesByChef()
