+++ src/js/pages/edit/editTypeOfFoodInRecipe.spec.js
@@ -1,12 +1,13 @@
 const { expect } = require("chai");
 const Ingredient = require("../../model/Ingredient");
 const Recipe = require("../../model/Recipe");
-const addOrEditIngredientInRecipe = require("./addOrEdditIngredientInRecipe");
+const editTypeOfFoodInRecipe = require("./editTypeOfFoodInRecipe");
 
 let ingredients;
 let steps;
+let typeOfFood;
 let recipe;
-describe("addOrEditIngredientInRecipe", () => {
+describe("editTypeOfFoodInRecipe", () => {
 	before(() => {
 		ingredients = [
 			{ ingredient: new Ingredient(1, "leche"), quantity: "200 g" },
@@ -15,23 +16,21 @@ describe("addOrEditIngredientInRecipe", () => {
 			{ ingredient: new Ingredient(4, "asuca"), quantity: "30g" },
 			{ ingredient: new Ingredient(5, "sumo naranja"), quantity: "400g" }];
 		steps = ["a", "b", "c", "d", "f", "g"];
+		typeOfFood = "Postre";
 		recipe = new Recipe(
 			1,
 			"poderosa",
 			ingredients,
 			steps,
-			"mexicana",
+			typeOfFood,
 			"china",
 			1
 		);
 	});
 	it("Debería devolvernos una receta", () => {
-		expect(addOrEditIngredientInRecipe(recipe, "hola")).to.be.an.instanceof(Recipe);
+		expect(editTypeOfFoodInRecipe(recipe, "Postre")).to.be.an.instanceof(Recipe);
 	});
-	it("Debería devolvernos una receta con 6 ingredientes", () => {
-		expect(addOrEditIngredientInRecipe(recipe, "hola").ingredients).that.have.length(6);
-	});
-	it("Debería devolvernos una receta con 5 ingredientes en la cual el 4 sea \"asuca\"", () => {
-		expect(addOrEditIngredientInRecipe(recipe, "asuca", 3).ingredients[3]).to.equal("asuca");
+	it("Debería devolvernos una receta con 7 pasos", () => {
+		expect(editTypeOfFoodInRecipe(recipe, "Postre").typeOfFood).to.equal("Postre");
 	});
 });
